+++ Dilgine/Dilgine/Scripts/Core/Vector3.h
@@ -5,11 +5,11 @@
 
 struct Vector3
 {
-	int x, y, z;
+	float x, y, z;
 
 	Vector3() { x = 0; y = 0; z = 0; }
-	Vector3(int val) { x = val; y = val; z = val; }
-	Vector3(int valX, int valY, int valZ) { x = valX; y = valY; z = valZ; }
+	Vector3(float val) { x = val; y = val; z = val; }
+	Vector3(float valX, float valY, float valZ) { x = valX; y = valY; z = valZ; }
 
 	float Magnitude() { return sqrt((x * x) + (y * y) + (z * z)); }
 	Vector3 Normalized() {
@@ -21,10 +21,10 @@ struct Vector3
 
 	Vector3 operator+(const Vector3& rhs) { return Vector3(x + rhs.x, y + rhs.y, z + rhs.z); }
 	Vector3 operator-(const Vector3& rhs) { return Vector3(x - rhs.x, y - rhs.y, z - rhs.z); }
-	Vector3 operator*(const int rhs) { return Vector3(x * rhs, y * rhs, z * rhs); }
+	Vector3 operator*(const float rhs) { return Vector3(x * rhs, y * rhs, z * rhs); }
 	void operator+=(const Vector3& rhs) { x += rhs.x; y += rhs.y; z += rhs.z; }
 	void operator-=(const Vector3& rhs) { x -= rhs.x; y -= rhs.y; z -= rhs.z; }
-	void operator*=(const int rhs) { x *= rhs; y *= rhs; z *= rhs; }
+	void operator*=(const float rhs) { x *= rhs; y *= rhs; z *= rhs; }
 
 	bool operator==(const Vector3& rhs) const { return (x == rhs.x && y == rhs.y && z == rhs.z); }
 	bool operator!=(const Vector3& rhs) const { return (x != rhs.x || y != rhs.y || z != rhs.z); }
